requirements.txt
pandas
python-docx
openpyxl
setuptools